requirements.txt
albumentations==1.3.1
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
configparser==6.0.0
contourpy==1.1.1
cycler==0.12.0
debugpy==1.8.0
decorator==5.1.1
Django==4.2.6
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework==3.14.0
executing==2.0.0
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.43.0
fsspec==2023.9.2
humanfriendly==10.0
idna==3.4
imageio==2.31.6
imutils==0.5.4
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.2
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.1
numpy==1.26.0
onnx==1.14.1
onnxruntime==1.16.0
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycocotools==2.0.7
Pygments==2.16.1
PyMySQL==1.1.0
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
qudida==0.0.4
requests==2.31.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
shutils==0.1.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.9.26
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.211
urllib3==2.0.6
wcwidth==0.2.8
